Use `internalChecksFilter` with `stabilityDays`
`stabilityDays` alone only adds a status check to a branch. In
combination with `prCreation="not-pending"`, this should result in PRs
only being created once X days have passed and the dependency is
considered 'stable'. However this behaviour doesn't work with groups as
explained here: https://github.com/renovatebot/config-help/issues/572

Additionally, the status check is mostly helpful when dealing with
automerging which we don't currently use. The original intention of
using `stabilityDays` was to ignore just published dependencies for
security and stability of new packages.

Using `internalChecksFilter` in tandem should achieve this original
goal. Now any 'pending' releases will be filtered, and PRs will be
skipped unless a non-pending version is available.